cleared Checkout component: - Moved out fieldList - Moved out generateBatch function
