+++ src/components/Checkout/formData.js
@@ -1,0 +1,23 @@
+
+export const fieldList = {
+    name : {
+        label : "Nombre",
+        type : "text",
+        ph : "Nombre Apellido"
+    },
+    email : {
+        label : "Tu correo",
+        type : "email",
+        ph : "usuario@correo.com"
+    },
+    email_val : {
+        label : "Tu correo nuevamente",
+        type : "email",
+        ph : "usuario@correo.com"
+    },
+    phone : {
+        label : "Tu teléfono",
+        type : "text",
+        ph : "#########"
+    },
+}
